update files structure
