+++ src/test/integration/services/accounts.e2e-spec.ts
@@ -1,401 +1,568 @@
-import { Test } from '@nestjs/testing';
-import { AccountService } from 'src/endpoints/accounts/account.service';
-import { PublicAppModule } from 'src/public.app.module';
-import { DeployedContract } from 'src/endpoints/accounts/entities/deployed.contract';
-import '@multiversx/sdk-nestjs-common/lib/utils/extensions/jest.extensions';
-import { ApiConfigService } from 'src/common/api-config/api.config.service';
 import { CacheService } from "@multiversx/sdk-nestjs-cache";
-import { AddressUtils } from '@multiversx/sdk-nestjs-common';
-import { ElasticService } from '@multiversx/sdk-nestjs-elastic';
-import { AccountKey } from 'src/endpoints/accounts/entities/account.key';
-import { AccountEsdtHistory } from 'src/endpoints/accounts/entities/account.esdt.history';
-import { AccountFilter } from 'src/endpoints/accounts/entities/account.filter';
-import { AccountHistoryFilter } from 'src/endpoints/accounts/entities/account.history.filter';
-import { Guardian } from 'src/common/gateway/entities/guardian';
-import { GuardianResult } from 'src/common/gateway/entities/guardian.result';
+import { AddressUtils } from "@multiversx/sdk-nestjs-common";
+import { ApiService } from "@multiversx/sdk-nestjs-http";
+import { Test } from "@nestjs/testing";
+import { ApiConfigService } from "src/common/api-config/api.config.service";
+import { AssetsService } from "src/common/assets/assets.service";
+import { GatewayService } from "src/common/gateway/gateway.service";
+import { IndexerService } from "src/common/indexer/indexer.service";
+import { PluginService } from "src/common/plugins/plugin.service";
+import { ProtocolService } from "src/common/protocol/protocol.service";
+import { AccountService } from "src/endpoints/accounts/account.service";
+import { AccountDetailed } from "src/endpoints/accounts/entities/account.detailed";
+import { AccountFilter } from "src/endpoints/accounts/entities/account.filter";
+import { AccountHistoryFilter } from "src/endpoints/accounts/entities/account.history.filter";
+import { SmartContractResultService } from "src/endpoints/sc-results/scresult.service";
+import { StakeService } from "src/endpoints/stake/stake.service";
+import { TransactionFilter } from "src/endpoints/transactions/entities/transaction.filter";
+import { TransactionType } from "src/endpoints/transactions/entities/transaction.type";
+import { TransactionService } from "src/endpoints/transactions/transaction.service";
+import { TransferService } from "src/endpoints/transfers/transfer.service";
+import { UsernameService } from "src/endpoints/usernames/username.service";
+import { VmQueryService } from "src/endpoints/vm.query/vm.query.service";
 
 describe('Account Service', () => {
-  let accountService: AccountService;
+  let service: AccountService;
+  let indexerService: IndexerService;
+  let cacheService: CacheService;
+  let apiService: ApiService;
+  let apiConfigService: ApiConfigService;
+  let transactionService: TransactionService;
+  let transferService: TransferService;
+  let smartContractResultService: SmartContractResultService;
 
-  const accountHistory = [
-    {
-      address: 'erd1ss6u80ruas2phpmr82r42xnkd6rxy40g9jl69frppl4qez9w2jpsqj8x97',
-      balance: '3074841073460000000',
-      timestamp: 1649147436,
-      isSender: true,
-    },
-    {
-      address: 'erd1ss6u80ruas2phpmr82r42xnkd6rxy40g9jl69frppl4qez9w2jpsqj8x97',
-      balance: '3075399483460000000',
-      timestamp: 1649147376,
-      isSender: true,
-    },
-  ];
-
-  beforeAll(async () => {
+  beforeEach((async () => {
     const moduleRef = await Test.createTestingModule({
-      imports: [PublicAppModule],
+      providers: [
+        AccountService,
+        {
+          provide: IndexerService,
+          useValue: {
+            getAccount: jest.fn(),
+            getAccounts: jest.fn(),
+            getAccountsCount: jest.fn(),
+            getAccountsForAddresses: jest.fn(),
+            getAccountContracts: jest.fn(),
+            getAccountContractsCount: jest.fn(),
+            getAccountHistory: jest.fn(),
+            getAccountTokenHistory: jest.fn(),
+            getAccountHistoryCount: jest.fn(),
+            getAccountTokenHistoryCount: jest.fn(),
+            getScDeploy: jest.fn(),
+            getTransaction: jest.fn(),
+          },
+        },
+        {
+          provide: GatewayService,
+          useValue: {
+            getGuardianData: jest.fn(),
+            getAddressDetails: jest.fn(),
+            getNetworkStatus: jest.fn(),
+          },
+        },
+        {
+          provide: CacheService,
+          useValue: {
+            getOrSet: jest.fn(),
+          },
+        },
+        {
+          provide: VmQueryService,
+          useValue: {
+            vmQuery: jest.fn(),
+          },
+        },
+        {
+          provide: ApiConfigService,
+          useValue: {
+            getVerifierUrl: jest.fn(),
+            getIsIndexerV3FlagActive: jest.fn(),
+            getDelegationContractAddress: jest.fn(),
+            getAuctionContractAddress: jest.fn(),
+            getStakingContractAddress: jest.fn(),
+          },
+        },
+        {
+          provide: TransactionService,
+          useValue: {
+            getTransactionCountForAddress: jest.fn(),
+          },
+        },
+        {
+          provide: PluginService,
+          useValue: {
+            processAccount: jest.fn(),
+          },
+        },
+        {
+          provide: StakeService,
+          useValue: {
+            getAllStakesForNode: jest.fn(),
+          },
+        },
+        {
+          provide: TransferService,
+          useValue: {
+            getTransfersCount: jest.fn(),
+          },
+        },
+        {
+          provide: SmartContractResultService,
+          useValue: {
+            getAccountScResultsCount: jest.fn(),
+          },
+        },
+        {
+          provide: AssetsService,
+          useValue: {
+            getAllAccountAssets: jest.fn(),
+          },
+        },
+        {
+          provide: UsernameService,
+          useValue: {
+            getUsernameForAddress: jest.fn(),
+          },
+        },
+        {
+          provide: ApiService,
+          useValue: {
+            get: jest.fn(),
+          },
+        },
+        {
+          provide: ProtocolService,
+          useValue: {
+            getShardCount: jest.fn(),
+          },
+        },
+      ],
     }).compile();
 
-    accountService = moduleRef.get<AccountService>(AccountService);
+    service = moduleRef.get<AccountService>(AccountService);
+    indexerService = moduleRef.get<IndexerService>(IndexerService);
+    cacheService = moduleRef.get<CacheService>(CacheService);
+    apiService = moduleRef.get<ApiService>(ApiService);
+    apiConfigService = moduleRef.get<ApiConfigService>(ApiConfigService);
+    transactionService = moduleRef.get<TransactionService>(TransactionService);
+    transferService = moduleRef.get<TransferService>(TransferService);
+    smartContractResultService = moduleRef.get<SmartContractResultService>(SmartContractResultService);
+  }));
+
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
-  beforeEach(() => { jest.restoreAllMocks(); });
+  it('service should be defined', () => {
+    expect(service).toBeDefined();
+  });
 
+  describe('getAccountsCount', () => {
+    it('should call cachingService.getOrSet if filter.ownerAddress is not provided', async () => {
+      const filter: AccountFilter = { ownerAddress: undefined };
+      const expectedResult = 5;
 
-  describe("getAccountsCount", () => {
-    it("should return total accounts count", async () => {
+      jest.spyOn(cacheService, 'getOrSet').mockResolvedValue(expectedResult);
+      jest.spyOn(indexerService, 'getAccountsCount').mockResolvedValue(expectedResult);
 
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
+      const result = await service.getAccountsCount(filter);
 
-      jest
-        .spyOn(ElasticService.prototype, 'getCount')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_address: string) => 49100));
+      expect(cacheService.getOrSet).toHaveBeenCalled();
+      expect(indexerService.getAccountsCount).not.toHaveBeenCalled();
+      expect(result).toEqual(expectedResult);
+    });
 
-      const results = await accountService.getAccountsCount(new AccountFilter());
+    it('should call indexerService.getAccountsCount directly if filter.ownerAddress is provided', async () => {
+      const filter = { ownerAddress: 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz' };
+      const expectedResult = 10;
 
-      expect(results).toStrictEqual(49100);
+      jest.spyOn(cacheService, 'getOrSet').mockResolvedValue(expectedResult);
+      jest.spyOn(indexerService, 'getAccountsCount').mockResolvedValue(expectedResult);
+
+      const result = await service.getAccountsCount(filter);
+
+      expect(cacheService.getOrSet).not.toHaveBeenCalled();
+      expect(indexerService.getAccountsCount).toHaveBeenCalled();
+      expect(result).toEqual(expectedResult);
     });
   });
 
-  describe.only("getAccount", () => {
-    it("should return null because test simulates that address is not valid ", async () => {
-      const mock_isAddressValid = jest.spyOn(AddressUtils, 'isAddressValid');
-      mock_isAddressValid.mockImplementation(() => false);
+  describe('getAccountVerification', () => {
+    it('should return null if the address is not valid', async () => {
+      const address = 'invalid_address';
 
-      const address: string = " ";
-      const results = await accountService.getAccount(address);
+      jest.spyOn(AddressUtils, 'isAddressValid').mockReturnValue(false);
 
-      expect(results).toBeNull();
+      const result = await service.getAccountVerification(address);
+
+      expect(result).toBeNull();
     });
 
-    it("should return account details", async () => {
-      const mock_isAddressValid = jest.spyOn(AddressUtils, 'isAddressValid');
-      mock_isAddressValid.mockImplementation(() => true);
+    it('should return verification data if the address is valid', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const verifierUrl = 'https://play-api.multiversx.com';
+      const verificationData = { verified: true };
 
-      const address: string = "erd1cnyng48s8lrjn95rpdfgykxl5993c5qhn5jqt0ar960f7v3umnrsy9yx0s";
-      const results = await accountService.getAccount(address);
+      jest.spyOn(AddressUtils, 'isAddressValid').mockReturnValue(true);
+      jest.spyOn(apiConfigService, 'getVerifierUrl').mockReturnValue(verifierUrl);
+      jest.spyOn(apiService, 'get').mockResolvedValue({ data: verificationData });
 
-      expect(results).toHaveProperties(
-        ['address', 'balance', 'nonce', 'shard', 'code',
-          'codeHash', 'rootHash', 'txCount', 'scrCount',
-          'username', 'shard', 'developerReward', 'ownerAddress', 'scamInfo',
-        ]);
+      const result = await service.getAccountVerification(address);
+
+      expect(apiService.get).toHaveBeenCalledWith(`${verifierUrl}/verifier/${address}`);
+      expect(result).toEqual(verificationData);
+    });
+  });
+
+  describe('getAccountSimple', () => {
+    it('should return null if the address is not valid', async () => {
+      const address = 'invalid_address';
+
+      jest.spyOn(AddressUtils, 'isAddressValid').mockReturnValue(false);
+
+      const result = await service.getAccountSimple(address);
+
+      expect(result).toBeNull();
     });
 
-    it('should return account details with isGuarded = true and guardian data extra fields when isGuarded is true in codeAttributes', async () => {
-      const address = 'erd1spyavw0956vq68xj8y4tenjpq2wd5a9p2c6j8gsz7ztyrnpxrruqzu66jx';
-
-      const mockGuardianData: GuardianResult = {
-        guardianData: {
-          activeGuardian: new Guardian({
-            activationEpoch: 9,
-            address: 'erd1spyavw0956vq68xj8y4tenjpq2wd5a9p2c6j8gsz7ztyrnpxrruqzu66jx',
-            serviceUID: 'ServiceID',
-          }),
-          pendingGuardian: new Guardian({
-            activationEpoch: 13,
-            address: 'erd1k2s324ww2g0yj38qn2ch2jwctdy8mnfxep94q9arncc6xecg3xaq6mjse8',
-            serviceUID: 'serviceUID',
-          }),
-          guarded: true,
-        },
+    it('should return account data if the address is valid', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const accountData: AccountDetailed = {
+        address: 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz',
+        balance: '162486906126924046',
+        nonce: 45,
+        timestamp: 0,
+        shard: 0,
+        ownerAddress: '',
+        assets: undefined,
+        code: '',
+        codeHash: "",
+        rootHash: 'w4fUiW+zHBmft9XlGbzVfcfn3rMtWKwi4bF+cjPPZ2k=',
+        txCount: 0,
+        scrCount: 0,
+        username: 'alice.elrond',
+        developerReward: '0',
+        isPayableBySmartContract: undefined,
+        scamInfo: undefined,
+        nftCollections: undefined,
+        nfts: undefined,
       };
 
-      const mock_decodeCodeMetadata = jest.spyOn(AddressUtils, 'decodeCodeMetadata');
-      mock_decodeCodeMetadata.mockImplementation((_codeMetadata: string) => {
-        return {
-          isUpgradeable: false,
-          isReadable: true,
-          isGuarded: true,
-          isPayable: false,
-          isPayableBySmartContract: false,
-        };
+      jest.spyOn(AddressUtils, 'isAddressValid').mockReturnValue(true);
+      jest.spyOn(service, 'getAccountRaw').mockResolvedValue(accountData);
+
+      const result = await service.getAccountSimple(address);
+
+      expect(service.getAccountRaw).toHaveBeenCalledWith(address);
+      expect(result).toEqual(accountData);
+    });
+  });
+
+  describe('getAccountTxCount', () => {
+    it('should return account transactions count from transaction service if indexer-v3 is false', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const expectedTxCount = 100;
+
+      jest.spyOn(apiConfigService, 'getIsIndexerV3FlagActive').mockReturnValue(false);
+      jest.spyOn(transactionService, 'getTransactionCountForAddress').mockResolvedValue(expectedTxCount);
+
+      const result = await service.getAccountTxCount(address);
+
+      expect(transactionService.getTransactionCountForAddress).toHaveBeenCalledWith(address);
+      expect(transferService.getTransfersCount).not.toHaveBeenCalled();
+      expect(result).toStrictEqual(expectedTxCount);
+    });
+
+    it('should return account transactions count from transfer service if indexer-v3 is true', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const expectedTxCount = 100;
+
+      jest.spyOn(apiConfigService, 'getIsIndexerV3FlagActive').mockReturnValue(true);
+      jest.spyOn(transferService, 'getTransfersCount').mockResolvedValue(expectedTxCount);
+
+      const result = await service.getAccountTxCount(address);
+
+      expect(transferService.getTransfersCount).toHaveBeenCalledWith(new TransactionFilter(
+        { address: address, functions: [], receivers: [], senders: [], type: TransactionType.Transaction }));
+
+      expect(transactionService.getTransactionCountForAddress).not.toHaveBeenCalled();
+      expect(result).toStrictEqual(expectedTxCount);
+    });
+  });
+
+  describe('getAccountScResults', () => {
+    it('should return account smart contract results from smartContractResult service if indexer-v3 is false',
+      async () => {
+        const address = "erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz";
+        const expectedTxCount = 100;
+
+        jest.spyOn(apiConfigService, 'getIsIndexerV3FlagActive').mockReturnValue(false);
+        jest.spyOn(smartContractResultService, 'getAccountScResultsCount').mockResolvedValue(expectedTxCount);
+
+        const result = await service.getAccountScResults(address);
+
+        expect(smartContractResultService.getAccountScResultsCount).toHaveBeenCalledWith(address);
+        expect(transferService.getTransfersCount).not.toHaveBeenCalled();
+        expect(result).toStrictEqual(expectedTxCount);
       });
 
-      jest.spyOn(accountService['gatewayService'], 'getGuardianData').mockResolvedValue(mockGuardianData);
-      const result = await accountService.getAccount(address, undefined, true);
+    it('should return account smart contract results from transfer service if indexer-v3 is true',
+      async () => {
+        const address = "erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz";
+        const expectedTxCount = 100;
 
-      expect(result?.isGuarded).toStrictEqual(true);
-      expect(result?.activeGuardianActivationEpoch).toStrictEqual(9);
-      expect(result?.activeGuardianAddress).toStrictEqual('erd1spyavw0956vq68xj8y4tenjpq2wd5a9p2c6j8gsz7ztyrnpxrruqzu66jx');
-      expect(result?.activeGuardianServiceUid).toStrictEqual('ServiceID');
-      expect(result?.pendingGuardianActivationEpoch).toStrictEqual(13);
-      expect(result?.pendingGuardianAddress).toStrictEqual('erd1k2s324ww2g0yj38qn2ch2jwctdy8mnfxep94q9arncc6xecg3xaq6mjse8');
-      expect(result?.pendingGuardianServiceUid).toStrictEqual('serviceUID');
+        jest.spyOn(apiConfigService, 'getIsIndexerV3FlagActive').mockReturnValue(true);
+        jest.spyOn(transferService, 'getTransfersCount').mockResolvedValue(expectedTxCount);
+
+        const result = await service.getAccountScResults(address);
+
+        expect(transferService.getTransfersCount).toHaveBeenCalledWith(new TransactionFilter(
+          { address: address, functions: [], receivers: [], senders: [], type: TransactionType.SmartContractResult }));
+
+        expect(smartContractResultService.getAccountScResultsCount).not.toHaveBeenCalledWith(address);
+        expect(result).toStrictEqual(expectedTxCount);
+      });
+  });
+
+  describe('getAccountDeployedAt', () => {
+    it('should call cachingService.getOrSet with the correct key and TTL', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const expectedResult = 1620000000;
+
+      jest.spyOn(cacheService, 'getOrSet').mockResolvedValue(expectedResult);
+      jest.spyOn(service, 'getAccountDeployedAtRaw').mockResolvedValue(expectedResult);
+
+      const result = await service.getAccountDeployedAt(address);
+
+      expect(cacheService.getOrSet).toHaveBeenCalled();
+      expect(result).toEqual(expectedResult);
+    });
+  });
+
+  describe('getAccountDeployedAtRaw', () => {
+    it('should return null if no scDeploy is found', async () => {
+      const address = 'erd1qqqqqqqqqqqqqpgqa0fsfshnff4n76jhcye6k7uvd7qacsq42jpsp6shh2';
+
+      jest.spyOn(indexerService, 'getScDeploy').mockResolvedValue(undefined);
+
+      const result = await service.getAccountDeployedAtRaw(address);
+
+      expect(indexerService.getScDeploy).toHaveBeenCalledWith(address);
+      expect(result).toEqual(null);
     });
 
-    it('should return account details with isGuarded = false when isGuarded is false in codeAttributes', async () => {
-      const address = 'erd1spyavw0956vq68xj8y4tenjpq2wd5a9p2c6j8gsz7ztyrnpxrruqzu66jx';
-      const mockGuardianData: GuardianResult = {
-        guardianData: {
-          activeGuardian: undefined,
-          pendingGuardian: undefined,
-          guarded: false,
-        },
-      };
+    it('should return null if no deployTxHash is found', async () => {
+      const address = 'erd1qqqqqqqqqqqqqpgqa0fsfshnff4n76jhcye6k7uvd7qacsq42jpsp6shh2';
 
-      const mock_decodeCodeMetadata = jest.spyOn(AddressUtils, 'decodeCodeMetadata');
-      mock_decodeCodeMetadata.mockImplementation((_codeMetadata: string) => {
-        return {
-          isUpgradeable: false,
-          isReadable: true,
-          isGuarded: false,
-          isPayable: false,
-          isPayableBySmartContract: false,
-        };
+      jest.spyOn(indexerService, 'getScDeploy').mockResolvedValue(
+        Promise.resolve({ deployTxHash: '', address: address, contract: address, deployer: address, timestamp: 1620000000, upgrades: [] }));
+
+      const result = await service.getAccountDeployedAtRaw(address);
+
+      expect(indexerService.getScDeploy).toHaveBeenCalledWith(address);
+      expect(result).toBeNull();
+    });
+
+    it('should return the timestamp if scDeploy, deployTxHash, and transaction are found', async () => {
+      const address = 'erd1qqqqqqqqqqqqqpgqa0fsfshnff4n76jhcye6k7uvd7qacsq42jpsp6shh2';
+      const txHash = "ca7acccc20a07695ba5657aac9c6e97b50fdcc9a77763447b9003721812271c7";
+      const timestamp = 1620000000;
+
+      jest.spyOn(indexerService, 'getScDeploy').mockResolvedValue(
+        Promise.resolve({ deployTxHash: txHash, address: address, contract: address, deployer: address, timestamp: 1620000000, upgrades: [] }));
+
+      jest.spyOn(indexerService, 'getTransaction').mockResolvedValue({
+        hash: 'ca7acccc20a07695ba5657aac9c6e97b50fdcc9a77763447b9003721812271c7',
+        miniBlockHash: '',
+        nonce: 100,
+        round: 100,
+        value: "100000",
+        receiver: '',
+        sender: '',
+        receiverShard: 1,
+        senderShard: 2,
+        gasPrice: '20000',
+        gasLimit: '20000',
+        gasUsed: '20000',
+        fee: '100',
+        data: '',
+        signature: '',
+        timestamp: 1620000000,
+        status: 'success',
+        searchOrder: 1,
+        hasScResults: true,
+        hasOperations: true,
+        tokens: [],
+        esdtValues: [],
+        receivers: [],
+        receiversShardIDs: [],
+        operation: '',
+        scResults: [],
       });
 
-      jest.spyOn(accountService['gatewayService'], 'getGuardianData').mockResolvedValue(mockGuardianData);
-      const result = await accountService.getAccount(address);
+      const result = await service.getAccountDeployedAtRaw(address);
 
-      expect(result?.isGuarded).toStrictEqual(false);
-      expect(result?.activeGuardianActivationEpoch).toBeUndefined();
-      expect(result?.activeGuardianAddress).toBeUndefined();
-      expect(result?.activeGuardianServiceUid).toBeUndefined();
-      expect(result?.pendingGuardianActivationEpoch).toBeUndefined();
-      expect(result?.pendingGuardianAddress).toBeUndefined();
-      expect(result?.pendingGuardianServiceUid).toBeUndefined();
+      expect(indexerService.getScDeploy).toHaveBeenCalledWith(address);
+      expect(indexerService.getTransaction).toHaveBeenCalledWith(txHash);
+      expect(result).toEqual(timestamp);
     });
 
-    it.skip("should return account details if IndexerV3Flag is active", async () => {
-      jest.spyOn(ApiConfigService.prototype, 'getIsIndexerV3FlagActive')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(() => true));
+    it('should return null if transaction is not found', async () => {
+      const address = 'erd1qqqqqqqqqqqqqpgqa0fsfshnff4n76jhcye6k7uvd7qacsq42jpsp6shh2';
+      const txHash = "invalid-tx";
 
-      const mock_isAddressValid = jest.spyOn(AddressUtils, 'isAddressValid');
-      mock_isAddressValid.mockImplementation(() => true);
+      jest.spyOn(indexerService, 'getScDeploy').mockResolvedValue(
+        Promise.resolve({ deployTxHash: txHash, address: address, contract: address, deployer: address, timestamp: 1620000000, upgrades: [] }));
+      jest.spyOn(indexerService, 'getTransaction').mockResolvedValue(null);
 
-      const address: string = "erd1cnyng48s8lrjn95rpdfgykxl5993c5qhn5jqt0ar960f7v3umnrsy9yx0s";
-      const results = await accountService.getAccount(address);
+      const result = await service.getAccountDeployedAtRaw(address);
 
-      expect(results).toHaveProperties(
-        ['address', 'balance', 'nonce', 'shard', 'code',
-          'codeHash', 'rootHash', 'txCount', 'scrCount',
-          'username', 'shard', 'developerReward', 'ownerAddress', 'scamInfo',
-        ]);
+      expect(indexerService.getScDeploy).toHaveBeenCalledWith(address);
+      expect(indexerService.getTransaction).toHaveBeenCalledWith(txHash);
+      expect(result).toBeNull();
     });
   });
 
-  describe("getAccountDeployedAt", () => {
-    it("should return the deployed timestamp for a given address", async () => {
+  describe('getAccountDeployedTxHash', () => {
+    it('should call cachingService.getOrSet with the correct key and TTL', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const expectedResult = '1620000000';
 
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
+      jest.spyOn(cacheService, 'getOrSet').mockResolvedValue(expectedResult);
+      jest.spyOn(service, 'getAccountDeployedTxHashRaw').mockResolvedValue(expectedResult);
 
-      jest
-        .spyOn(AccountService.prototype, 'getAccountDeployedAtRaw')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_address: string) => 1616769300));
+      const result = await service.getAccountDeployedTxHash(address);
 
-      const address: string = "erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq8hlllls7a6h85";
-      const results = await accountService.getAccountDeployedAt(address);
-
-      expect(results).toStrictEqual(1616769300);
-    });
-
-    it("should return null because test simulates that scDeployed is undefined", async () => {
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
-
-      jest
-        .spyOn(ElasticService.prototype, 'getItem')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_collection: string, _key: string, _identifier: string) => undefined));
-
-      jest
-        .spyOn(AccountService.prototype, 'getAccountDeployedAtRaw')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_address: string) => null));
-
-      const address: string = "erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq8hlllls7a6h85";
-      const results = await accountService.getAccountDeployedAt(address);
-
-      expect(results).toBeNull();
+      expect(cacheService.getOrSet).toHaveBeenCalled();
+      expect(result).toEqual(expectedResult);
     });
   });
 
-  describe("getAccountDeployedAtRaw", () => {
-    it("should return account deployed timestamp because test simulates that account is a smart-contract", async () => {
-      const address: string = "erd1qqqqqqqqqqqqqpgqvc7gdl0p4s97guh498wgz75k8sav6sjfjlwqh679jy";
-      const results = await accountService.getAccountDeployedAtRaw(address);
+  describe('getAccountDeployedTxHashRaw', () => {
+    const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+    const txHash = 'ca7acccc20a07695ba5657aac9c6e97b50fdcc9a77763447b9003721812271c7';
 
-      expect(results).toStrictEqual(1636897470);
+    it('should return null if no scDeploy is found', async () => {
+      jest.spyOn(indexerService, 'getScDeploy').mockResolvedValue(undefined);
+
+      const result = await service.getAccountDeployedTxHashRaw(address);
+
+      expect(indexerService.getScDeploy).toHaveBeenCalledWith(address);
+      expect(result).toBeNull();
     });
 
-    it("should return null because test simulates that scDeployed is undefined and should return null", async () => {
-      jest
-        .spyOn(ElasticService.prototype, 'getItem')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_collection: string, _key: string, _identifier: string) => undefined));
+    it('should return account deployed txhash value', async () => {
+      jest.spyOn(indexerService, 'getScDeploy').mockResolvedValue(
+        Promise.resolve({
+          deployTxHash: txHash,
+          address: address, contract: address, deployer: address, timestamp: 1620000000, upgrades: [],
+        }));
 
-      const address: string = "erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq8hlllls7a6h85";
-      const results = await accountService.getAccountDeployedAtRaw(address);
+      const result = await service.getAccountDeployedTxHashRaw(address);
 
-      expect(results).toBeNull();
-    });
-  });
-
-  describe("getAccounts", () => {
-    it("should return 10 accounts", async () => {
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
-
-      const results = await accountService.getAccounts({ from: 0, size: 10 }, new AccountFilter());
-
-      expect(results).toHaveLength(10);
+      expect(indexerService.getScDeploy).toHaveBeenCalledWith(address);
+      expect(result).toStrictEqual(txHash);
     });
   });
 
-  describe("getDeferredAccount", () => {
-    it("should return empty list because test simulates that address is not deferred", async () => {
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
+  describe('getAccountIsVerified', () => {
+    const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+    const codeHash = 'ca7acccc20a07695ba5657aac9c6e97b50fdcc9a77763447b9003721812271c7';
 
-      const address: string = "erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz";
-      const results = await accountService.getDeferredAccount(address);
+    it('should call cachingService.getOrSet', async () => {
+      const getAccountIsVerifiedRawSpy = jest.spyOn(service, 'getAccountIsVerifiedRaw').mockResolvedValue(true);
 
-      expect(results).toStrictEqual([]);
+      // eslint-disable-next-line require-await
+      jest.spyOn(cacheService, 'getOrSet').mockImplementation(async (_, callback) => callback());
+
+      await service.getAccountIsVerified(address, codeHash);
+
+      expect(cacheService.getOrSet).toHaveBeenCalled();
+      expect(getAccountIsVerifiedRawSpy).toHaveBeenCalledWith(address, codeHash);
     });
   });
 
-  //KindReminder
-  describe("getKeys", () => {
-    it("should return a list of keys of type AccountKey for a specific address", async () => {
-      const mock_isAddressValid = jest.spyOn(AddressUtils, 'bech32Decode');
-      mock_isAddressValid.mockImplementation(() => "000000000000000000010000000000000000000000000000000000001effffff");
+  describe('getAccountIsVerifiedRaw', () => {
+    const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+    const codeHash = 'ca7acccc20a07695ba5657aac9c6e97b50fdcc9a77763447b9003721812271c7';
+    const verifierUrl = 'https://play-api.multiversx.com';
 
-      const results = await accountService.getKeys("erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq8hlllls7a6h85");
-
-      for (const result of results) {
-        expect(result).toHaveStructure(Object.keys(new AccountKey()));
-        expect(result).toHaveProperties(['blsKey', 'stake', 'topUp', 'status', 'rewardAddress', 'queueIndex', 'queueSize']);
-      }
+    beforeEach(() => {
+      jest.spyOn(apiConfigService, 'getVerifierUrl').mockReturnValue(verifierUrl);
     });
 
-    it("should return an empty list because test simulates that provider account does not contain any bls key", async () => {
-      const mock_isAddressValid = jest.spyOn(AddressUtils, 'bech32Decode');
-      mock_isAddressValid.mockImplementation(() => "");
+    it('should return true if codeHash matches', async () => {
+      const data = { codeHash: Buffer.from(codeHash, 'base64').toString('hex') };
+      jest.spyOn(apiService, 'get').mockResolvedValue({ data });
 
-      const results = await accountService.getKeys("erd1qqqqqqqqqqqqqqqpqqqqqqqqqqqqqqqqqqqqqqqqqqqqq8hlllls7a6h85");
+      const result = await service.getAccountIsVerifiedRaw(address, codeHash);
 
-      expect(results).toStrictEqual([]);
+      expect(apiService.get).toHaveBeenCalledWith(`${verifierUrl}/verifier/${address}/codehash`, undefined, expect.any(Function));
+      expect(result).toBe(true);
     });
-  });
 
-  describe("getAccountContracts", () => {
-    it("should return account contracts details", async () => {
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
+    it('should return null if codeHash does not match', async () => {
+      const data = { codeHash: 'differentCodeHash' };
+      jest.spyOn(apiService, 'get').mockResolvedValue({ data });
 
-      const address: string = "erd1ss6u80ruas2phpmr82r42xnkd6rxy40g9jl69frppl4qez9w2jpsqj8x97";
-      const results = await accountService.getAccountContracts({ from: 0, size: 2 }, address);
+      const result = await service.getAccountIsVerifiedRaw(address, codeHash);
 
-      expect(results).toHaveLength(2);
+      expect(apiService.get).toHaveBeenCalledWith(`${verifierUrl}/verifier/${address}/codehash`, undefined, expect.any(Function));
+      expect(result).toBe(null);
+    });
 
-      for (const result of results) {
-        expect(results).toHaveLength(2);
-        expect(result).toHaveStructure(Object.keys(new DeployedContract()));
-      }
+    it('should return null if an error occurs', async () => {
+      jest.spyOn(apiService, 'get').mockRejectedValue(new Error('Test error'));
+
+      const result = await service.getAccountIsVerifiedRaw(address, codeHash);
+
+      expect(apiService.get).toHaveBeenCalledWith(`${verifierUrl}/verifier/${address}/codehash`, undefined, expect.any(Function));
+      expect(result).toBe(null);
     });
   });
 
-  describe("getAccountContractsCount", () => {
-    it("should return total contracts count for a specific account address", async () => {
+  describe('getAccountContractsCount', () => {
+    it('should return the number of contracts', async () => {
+      const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+      const contractsCount = 5;
 
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
+      jest.spyOn(indexerService, 'getAccountContractsCount').mockResolvedValue(contractsCount);
 
-      jest
-        .spyOn(ElasticService.prototype, 'getCount')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async () => 20));
+      const result = await service.getAccountContractsCount(address);
 
-      const address: string = "erd1ss6u80ruas2phpmr82r42xnkd6rxy40g9jl69frppl4qez9w2jpsqj8x97";
-      const results = await accountService.getAccountContractsCount(address);
-
-      expect(results).toStrictEqual(20);
+      expect(indexerService.getAccountContractsCount).toHaveBeenCalledWith(address);
+      expect(result).toEqual(contractsCount);
     });
   });
 
-  describe("getAccountHistory", () => {
-    it("should return total contracts count for a specific account address", async () => {
+  describe('getAccountHistoryCount', () => {
+    const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+    const filter = new AccountHistoryFilter({});
+    const historyCount = 10;
 
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
+    it('should return the account history count', async () => {
+      jest.spyOn(indexerService, 'getAccountHistoryCount').mockResolvedValue(historyCount);
 
-      jest
-        .spyOn(ElasticService.prototype, 'getList')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async () => accountHistory));
+      const result = await service.getAccountHistoryCount(address, filter);
 
-      const address: string = "erd1ss6u80ruas2phpmr82r42xnkd6rxy40g9jl69frppl4qez9w2jpsqj8x97";
-      const results = await accountService.getAccountHistory(address, { from: 0, size: 2 }, new AccountHistoryFilter({}));
-
-      expect(results).toHaveLength(2);
-
-      expect(results).toEqual(expect.arrayContaining([
-        expect.objectContaining(
-          {
-            address: 'erd1ss6u80ruas2phpmr82r42xnkd6rxy40g9jl69frppl4qez9w2jpsqj8x97',
-            balance: '3074841073460000000',
-            timestamp: 1649147436,
-            isSender: true,
-          }
-        ),
-      ]));
+      expect(indexerService.getAccountHistoryCount).toHaveBeenCalledWith(address, filter);
+      expect(result).toEqual(historyCount);
     });
   });
 
-  describe("getAccountTokenHistory", () => {
-    it("should return account token history", async () => {
-      const token: string = "RIDE-7d18e9";
-      const address: string = "erd19w6f7jqnf4nqrdmq0m548crrc4v3dmrxtn7u3dngep2r078v30aqzzu6nc";
-      const results = await accountService.getAccountTokenHistory(address, token, { from: 0, size: 1 }, new AccountHistoryFilter({}));
+  describe('getAccountTokenHistoryCount', () => {
+    const address = 'erd1qga7ze0l03chfgru0a32wxqf2226nzrxnyhzer9lmudqhjgy7ycqjjyknz';
+    const tokenIdentifier = 'WEGLD-bd4d79';
+    const filter = new AccountHistoryFilter({});
+    const tokenHistoryCount = 5;
 
-      if (!results) {
-        throw new Error('Properties are not defined');
-      }
+    it('should return the account token history count', async () => {
+      jest.spyOn(indexerService, 'getAccountTokenHistoryCount').mockResolvedValue(tokenHistoryCount);
 
-      for (const result of results) {
-        expect(result).toHaveStructure(Object.keys(new AccountEsdtHistory()));
-      }
-    });
+      const result = await service.getAccountTokenHistoryCount(address, tokenIdentifier, filter);
 
-    it("should return empty list because test simulates that token is not defined/found", async () => {
-      jest
-        .spyOn(CacheService.prototype, 'getOrSet')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async (_key: string, promise: any) => promise()));
-
-      jest
-        .spyOn(ElasticService.prototype, 'getList')
-        // eslint-disable-next-line require-await
-        .mockImplementation(jest.fn(async () => []));
-
-      const token: string = "";
-      const address: string = "erd19w6f7jqnf4nqrdmq0m548crrc4v3dmrxtn7u3dngep2r078v30aqzzu6nc";
-      const results = await accountService.getAccountTokenHistory(address, token, { from: 0, size: 1 }, new AccountHistoryFilter({}));
-
-      expect(results).toStrictEqual([]);
+      expect(indexerService.getAccountTokenHistoryCount).toHaveBeenCalledWith(address, tokenIdentifier, filter);
+      expect(result).toEqual(tokenHistoryCount);
     });
   });
 });
